Move Watcher to its own file
